AI_Tutor/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI & LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.7.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
mysql-connector-python==8.2.0

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue & Caching (Optional)
redis==5.0.1
celery==5.3.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pyyaml==6.0.1

# MCP (Model Context Protocol) - 可选,用于接入外部工具服务器
# mcp>=1.0.0

# Web Search (使用httpx直接调用API，无需额外依赖)

# DashScope - 阿里云百炼云端TTS服务
dashscope>=1.14.0

# Edge TTS - 微软免费TTS（无需API Key）
edge-tts>=6.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0
PyMuPDF==1.23.8
